+++ run/models/explorer.js
@@ -437,7 +437,7 @@ module.exports = (sequelize, DataTypes) => {
     }
 
     async safeUpdateSubscription(stripePlanId, stripeId, cycleEndsAt, status) {
-        if (!stripePlanId || !stripeId || !cycleEndsAt || !status) throw new Error('Missing parameter');
+        if (!stripePlanId || !cycleEndsAt || !status) throw new Error('Missing parameter');
 
         if (sequelize.models.StripeSubscription.rawAttributes.status.values.indexOf(status) == -1)
             throw new Error('Invalid subscription status');
@@ -466,7 +466,7 @@ module.exports = (sequelize, DataTypes) => {
     }
 
     async safeUpdateSettings(settings) {
-        const ALLOWED_SETTINGS = ['name', 'slug', 'token', 'totalSupply', 'l1Explorer'];
+        const ALLOWED_SETTINGS = ['name', 'slug', 'token', 'totalSupply', 'l1Explorer', 'rpcServer'];
 
         const filteredSettings = {};
         Object.keys(settings).forEach(key => {