+++ tests/unit/components/ExplorerPlanSelector.spec.js
@@ -1,7 +1,8 @@
-import flushPromises from 'flush-promises';
-
 import ExplorerPlanSelector from '@/components/ExplorerPlanSelector.vue';
-const plans = [{ slug: '50', price: 10 }, { slug: '100', price: 100 }];
+const plans = [
+    { slug: '50', price: 10, capabilities: { skipBilling: false } },
+    { slug: '100', price: 100, capabilities: { skipBilling: false } }
+];
 
 beforeAll(() => {
     vi.stubGlobal('location', { assign: vi.fn() });
@@ -13,7 +14,6 @@ afterAll(() => {
 });
 
 describe('ExplorerPlanSelector.vue', () => {
-
     beforeEach(() => {
         vi.spyOn(server, 'getExplorerPlans').mockResolvedValue({ data: plans });
     });
@@ -24,15 +24,12 @@ describe('ExplorerPlanSelector.vue', () => {
                 stubs: ['Explorer-Plan-Card']
             }
         });
-
         await flushPromises();
-
         expect(wrapper.html()).toMatchSnapshot();
     });
 
     it('Should start a crypto subscription', async () => {
         vi.spyOn(server, 'startCryptoSubscription').mockResolvedValueOnce();
-
         const wrapper = mount(ExplorerPlanSelector, {
             props: {
                 explorerId: 1
@@ -49,17 +46,13 @@ describe('ExplorerPlanSelector.vue', () => {
                 })],
             }
         });
-
         await flushPromises();
-
         wrapper.vm.onPlanSelected('50');
-
         expect(server.startCryptoSubscription).toHaveBeenCalledWith('50', 1);
     });
 
     it('Should redirect to a stripe checkout url', async () => {
         vi.spyOn(server, 'createStripeExplorerCheckoutSession').mockResolvedValueOnce({ data: { url: 'stripe.com' }});
-
         const wrapper = mount(ExplorerPlanSelector, {
             props: {
                 explorerId: 1
@@ -76,18 +69,14 @@ describe('ExplorerPlanSelector.vue', () => {
                 })],
             }
         });
-
         await flushPromises();
-
         wrapper.vm.onPlanSelected('50');
-
         expect(server.createStripeExplorerCheckoutSession)
             .toHaveBeenCalledWith(1, '50', 'http://app.ethernal.local:8080/explorers/1?justCreated=true', 'http://app.ethernal.local:8080/explorers/1');
     });
 
     it('Should update a subscription', async () => {
         vi.spyOn(server, 'updateExplorerSubscription').mockResolvedValueOnce();
-
         const wrapper = mount(ExplorerPlanSelector, {
             global: {
                 stubs: ['Explorer-Plan-Card'],
@@ -107,16 +96,13 @@ describe('ExplorerPlanSelector.vue', () => {
             },
         });
         await flushPromises();
-
         wrapper.vm.onPlanSelected('50');
-
         expect(server.updateExplorerSubscription).toHaveBeenCalledWith(1, '50');
         expect(window.confirm).toHaveBeenCalledWith(expect.stringContaining('This plan is cheaper'));
     });
 
     it('Should cancel a subscription', async () => {
         vi.spyOn(server, 'cancelExplorerSubscription').mockResolvedValueOnce();
-
         const wrapper = mount(ExplorerPlanSelector, {
             global: {
                 stubs: ['Explorer-Plan-Card'],
@@ -135,17 +121,14 @@ describe('ExplorerPlanSelector.vue', () => {
                 pendingCancelation: false,
             },
         });
-
-        wrapper.vm.onPlanSelected()
         await flushPromises();
-
+        wrapper.vm.onPlanSelected()
         expect(server.cancelExplorerSubscription).toHaveBeenCalledWith(1);
         expect(window.confirm).toHaveBeenCalledWith(expect.stringContaining('If you cancel now'));
     });
 
     it('Should start a trial', async () => {
         vi.spyOn(server, 'startTrial').mockResolvedValueOnce();
-
         const wrapper = mount(ExplorerPlanSelector, {
             global: {
                 stubs: ['Explorer-Plan-Card'],
@@ -165,9 +148,7 @@ describe('ExplorerPlanSelector.vue', () => {
             }
         });
         await flushPromises();
-
         wrapper.vm.onPlanSelected('100')
-
         expect(server.startTrial).toHaveBeenCalledWith(1, '100');
     });
 });